♻️ Move developer API docs out of settings
