+++ backend/islands/apps/remotes/src/components/remotes/DeveloperApiDocs/index.ts
@@ -1,0 +1,1 @@
+export { default } from './DeveloperApiDocs';
