+++ backend/islands/apps/remotes/src/components/remotes/SettingsApp/pages/DeveloperApiSetting/index.ts
@@ -1,2 +1,1 @@
 export { default } from './DeveloperApiSetting';
-export { DeveloperApiReferenceDetail } from './DeveloperApiSetting';
